start with empty shared memory graph db implementation

+++ src/graphlab/database/basic_types.hpp
@@ -6,7 +6,8 @@ namespace graphlab {
 
 /// \ingroup group_graph_database
 /// identifies a vertex id
-typedef __uint128_t graph_vid_t;
+// typedef __uint128_t graph_vid_t;
+typedef uint64_t graph_vid_t;
 
 /// \ingroup group_graph_database
 /// integer field data type
